chore(launchdarkly-adapter): to use @flopflip/types
