+++ packages/launchdarkly-adapter/modules/index.js
@@ -1,3 +1,2 @@
 export default from './adapter';
-export { types } from './adapter';
 export { version } from '../package.json';
